+++ apps/scan/src/main/scala/org/lfdecentralizedtrust/splice/scan/store/ScanStore.scala
@@ -186,6 +186,15 @@ trait ScanStore
     ]
   ]]
 
+  def getValidatorLicenseByValidator(
+      validator: Vector[PartyId]
+  )(implicit tc: TraceContext): Future[Seq[
+    Contract[
+      splice.validatorlicense.ValidatorLicense.ContractId,
+      splice.validatorlicense.ValidatorLicense,
+    ]
+  ]]
+
   def getBaseRateTrafficLimitsAsOf(t: CantonTimestamp)(implicit
       tc: TraceContext
   ): Future[splice.decentralizedsynchronizer.BaseRateTrafficLimits] =
@@ -384,6 +393,7 @@ object ScanStore {
             ScanAcsStoreRowData(
               contract = contract,
               validatorLicenseRoundsCollected = Some(roundsCollected.orElse(0L)),
+              validator = Some(PartyId.tryFromProtoPrimitive(contract.payload.validator)),
             )
         },
         mkFilter(splice.dso.svstate.SvNodeState.COMPANION)(co => co.payload.dso == dso) {